Moved json serializer to correct project directory
