+++ packages/langium/src/default-module.ts
@@ -8,7 +8,7 @@ import { DefaultScopeComputation, DefaultScopeProvider } from './references/scop
 import { DefaultNameProvider } from './references/naming';
 import { ValidationRegistry } from './service/validation/validation-registry';
 import { DefaultDocumentValidator } from './service/validation/document-validator';
-import { DefaultJsonSerializer } from './service/json-serializer/ast-json-serializer';
+import { DefaultJsonSerializer } from './serializer/json-serializer';
 import { DefaultCompletionProvider } from './service/completion/completion-provider';
 import { RuleInterpreter } from './service/completion/rule-interpreter';
